Fix two query bugs

+++ libsrc/geomodelgrids/serial/Block.cc
@@ -165,7 +165,7 @@ geomodelgrids::serial::Block::setHyperslabDims(const size_t dims[],
     } // if
     assert(dims);
 
-    for (int i = 0; i < ndims; ++i) {
+    for (size_t i = 0; i < ndims; ++i) {
         _hyperslabDims[i] = dims[i];
     } // for
 } // setHyperslabDims
